requirements.txt
tesseract==0.1.3
pytesseract==0.3.10
opencv-python==4.7.0.72
Pillow==9.4.0